# account-service/src/main/resources/openapi/account-api.yaml
openapi: 3.0.3
info:
  title: Account Service API
  description: API for managing accounts, movements and reports
  version: 1.0.0

servers:
  - url: http://localhost:8081

tags:
  - name: Account
    description: Account management API
  - name: Movement
    description: Movement management API
  - name: Report
    description: Report generation API

paths:

  # =====================================================
  # ACCOUNT
  # =====================================================

  /accounts:

    post:
      tags:
        - Account
      summary: Create a new account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequestDTO'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDTO'
        '400':
          description: Invalid input
        '404':
          description: Customer not found

    get:
      tags:
        - Account
      summary: Get all accounts
      operationId: getAllAccounts
      responses:
        '200':
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponseDTO'

  /accounts/{id}:

    get:
      tags:
        - Account
      summary: Get account by ID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDTO'
        '404':
          description: Account not found

    put:
      tags:
        - Account
      summary: Update account by ID
      operationId: updateAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequestDTO'
      responses:
        '200':
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDTO'
        '400':
          description: Invalid input
        '404':
          description: Account not found

    delete:
      tags:
        - Account
      summary: Delete account by ID
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Account deleted successfully
        '404':
          description: Account not found


  # =====================================================
  # MOVEMENT
  # =====================================================

  /movements:

    post:
      tags:
        - Movement
      summary: Register a new movement
      operationId: registerMovement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementRequestDTO'
      responses:
        '201':
          description: Movement registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementResponseDTO'
        '400':
          description: Invalid input or insufficient balance
        '404':
          description: Account not found

    get:
      tags:
        - Movement
      summary: Get all movements
      operationId: getAllMovements
      responses:
        '200':
          description: Movements retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovementResponseDTO'

  /movements/{id}:

    get:
      tags:
        - Movement
      summary: Get movement by ID
      operationId: getMovementById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Movement found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementResponseDTO'
        '404':
          description: Movement not found

    put:
      tags:
        - Movement
      summary: Update movement by ID
      operationId: updateMovement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementRequestDTO'
      responses:
        '200':
          description: Movement updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementResponseDTO'
        '400':
          description: Invalid input
        '404':
          description: Movement not found

    delete:
      tags:
        - Movement
      summary: Delete movement by ID
      operationId: deleteMovement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Movement deleted successfully
        '404':
          description: Movement not found


  # =====================================================
  # REPORT
  # =====================================================

  /reports/{customerId}:

    get:
      tags:
        - Report
      summary: Generate customer report with account movements
      operationId: generateReport
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64

        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date

        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date

      responses:
        '200':
          description: Report generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerReportDTO'
        '404':
          description: Customer or accounts not found


components:

  schemas:

    AccountRequestDTO:
      type: object
      required:
        - customerId
        - number
        - type
        - initialBalance
      properties:
        customerId:
          type: integer
          format: int64
        number:
          type: string
        type:
          type: string
        initialBalance:
          type: number
          format: double
        status:
          type: boolean

    AccountResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        number:
          type: string
        type:
          type: string
        balance:
          type: number
          format: double
        status:
          type: boolean

    MovementRequestDTO:
      type: object
      required:
        - accountId
        - type
        - value
      properties:
        accountId:
          type: integer
          format: int64
        type:
          type: string
        value:
          type: number
          format: double

    MovementResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        type:
          type: string
        value:
          type: number
          format: double
        balance:
          type: number
          format: double

    AccountReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: string
        type:
          type: string
        balance:
          type: number
          format: double
        status:
          type: boolean
        movements:
          type: array
          items:
            $ref: '#/components/schemas/MovementResponseDTO'

    CustomerReportDTO:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountReportDTO'